# backend/openapi.json
{
  "openapi": "3.0.3",
  "info": {"title":"NSSMS API","version":"1.0.0","description":"Versioned API for the National School Sports Management System."},
  "servers":[{"url":"http://localhost:3000"}],
  "components":{"securitySchemes":{"bearerAuth":{"type":"http","scheme":"bearer"}}},
  "paths":{
...
    "/api/v1/public/seasons/{id}":{"get":{"summary":"Read an approved public season","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"Season detail"}}}},
    "/api/v1/public/competitions/{id}":{"get":{"summary":"Read an approved public competition","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"Competition detail"}}}},
    "/api/v1/admin/announcements":{"get":{"security":[{"bearerAuth":[]}],"summary":"List announcements","responses":{"200":{"description":"Announcement list"}}},"post":{"security":[{"bearerAuth":[]}],"summary":"Publish an announcement","responses":{"201":{"description":"Created"}}}},
    "/api/v1/admin/entries":{"get":{"security":[{"bearerAuth":[]}],"summary":"List competition entries","responses":{"200":{"description":"Entry list"}}},"post":{"security":[{"bearerAuth":[]}],"summary":"Register a participant in a competition","responses":{"201":{"description":"Created"}}}},
...
    "/api/v1/admin/licenses/applications":{"post":{"security":[{"bearerAuth":[]}],"summary":"Create a license application","responses":{"201":{"description":"Created"}}}},
...
  }
}